prov/util: Add interface to get node name
- Move sock_getnodename to util as ofi_getnodename()
