+++ prov/sockets/include/sock.h
@@ -956,7 +956,6 @@ int sock_msg_verify_ep_attr(struct fi_ep_attr *ep_attr, struct fi_tx_attr *tx_at
 int sock_get_src_addr(struct sockaddr_in *dest_addr,
 		      struct sockaddr_in *src_addr);
 int sock_get_src_addr_from_hostname(struct sockaddr_in *src_addr, const char *service);
-void sock_getnodename(char *buf, int buflen);
 
 struct fi_info *sock_fi_info(enum fi_ep_type ep_type,
 			     struct fi_info *hints, void *src_addr, void *dest_addr);
