+++ prov/util/src/util_av.c
@@ -30,6 +30,8 @@
  * SOFTWARE.
  */
 
+#include "config.h"
+
 #include <arpa/inet.h>
 #include <ctype.h>
 #include <stdlib.h>
@@ -38,6 +40,11 @@
 #include <sys/types.h>
 #include <netdb.h>
 #include <netinet/in.h>
+
+#if HAVE_GETIFADDRS
+#include <net/if.h>
+#include <ifaddrs.h>
+#endif
 
 #include <fi_util.h>
 
@@ -92,6 +99,45 @@ out:
 	close(sock);
 	return ret;
 
+}
+
+void ofi_getnodename(char *buf, int buflen)
+{
+	int ret;
+	struct addrinfo ai, *rai = NULL;
+	struct ifaddrs *ifaddrs, *ifa;
+
+	ret = gethostname(buf, buflen);
+	if (ret == 0) {
+		memset(&ai, 0, sizeof(ai));
+		ai.ai_family = AF_INET;
+		ret = getaddrinfo(buf, NULL, &ai, &rai);
+		if (!ret) {
+			freeaddrinfo(rai);
+			return;
+		}
+	}
+
+#if HAVE_GETIFADDRS
+	ret = getifaddrs(&ifaddrs);
+	if (!ret) {
+		for (ifa = ifaddrs; ifa != NULL; ifa = ifa->ifa_next) {
+			if (ifa->ifa_addr == NULL || !(ifa->ifa_flags & IFF_UP) ||
+			     (ifa->ifa_addr->sa_family != AF_INET))
+				continue;
+
+			ret = getnameinfo(ifa->ifa_addr, sizeof(struct sockaddr_in),
+				  	  buf, buflen, NULL, 0, NI_NUMERICHOST);
+			if (ret == 0) {
+				freeifaddrs(ifaddrs);
+				return;
+			}
+		}
+		freeifaddrs(ifaddrs);
+	}
+#endif
+	/* no reasonable address found, try loopback */
+	strncpy(buf, "127.0.0.1", buflen);
 }
 
 int ofi_get_src_addr(uint32_t addr_format,
